% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brint}
\alias{brint}
\title{fully materialized MAE with all samples/features in intersection of iTRAQ, RNA-seq and RPPA for BRCA}
\format{MultiAssayExperiment}
\usage{
brint
}
\description{
fully materialized MAE with all samples/features in intersection of iTRAQ, RNA-seq and RPPA for BRCA
}
\note{
derived from brcaProteoMAE
}
\keyword{datasets}
