% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brProteoMAE}
\alias{brProteoMAE}
\title{partly delayed MAE with breast cancer pancan atlas references
and proteogenomics data from PMID 27251275 (Mertens et al Nature 2016)}
\format{MultiAssayExperiment}
\usage{
brProteoMAE
}
\description{
partly delayed MAE with breast cancer pancan atlas references
and proteogenomics data from PMID 27251275 (Mertens et al Nature 2016)
}
\note{
Assay retrieval required BigQuery authentication
}
\keyword{datasets}
