% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgconsis.R
\name{pgconsis}
\alias{pgconsis}
\title{define a shiny app to check consistency of iTRAQ, RNA-seq and RPPA
in BRCA}
\usage{
pgconsis()
}
\description{
define a shiny app to check consistency of iTRAQ, RNA-seq and RPPA
in BRCA
}
