% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3.R
\name{plot3}
\alias{plot3}
\title{a simple interactive display for iTRAQ vs RPPA vs RNA-seq}
\usage{
plot3(gene = "BCL2", tx = "none")
}
\arguments{
\item{gene}{character(1)}

\item{tx}{character(1) if 'none', raw data are displayed, otherwise z-scores (over tumors) are used}
}
\value{
a ggplot instance with a horizontal grid
}
\description{
a simple interactive display for iTRAQ vs RPPA vs RNA-seq
}
